+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.1
+
+- Added **PasswordStatus** class
+- Added **onPasswordChanged** method for direct textfield usage
+- Added **onPasswordStatusChange** method for streaming password status change
+
 ## 2.0.0
 - Updated the sdk.
 - Moved to dart from platform specific codes.